chore: changed subscriptions edit to raw flex divs. next convert to flexbox
