+++ src/writeData/subscriptions/components/BrokerDetails.scss
@@ -1,7 +1,6 @@
 @import '@influxdata/clockface/dist/variables.scss';
 .update-broker-form {
   float: right;
-  width: 75%;
 
   &__fixed {
     position: fixed;
