+++ src/writeData/subscriptions/components/ParsingDetails.scss
@@ -1,14 +1,14 @@
 @import '@influxdata/clockface/dist/variables.scss';
 .update-parsing-form {
   float: right;
-  width: 75%;
+  // width: 75%;
   .cf-form--wrapper {
     padding-bottom: 50px;
   }
 
   &--line-protocol {
     float: right;
-    width: 75%;
+    // width: 75%;
     .cf-form--wrapper {
       padding-bottom: 550px;
     }
