+++ src/writeData/subscriptions/components/SubscriptionDetails.scss
@@ -1,7 +1,7 @@
 @import '@influxdata/clockface/dist/variables.scss';
 .update-subscription-form {
   float: right;
-  width: 75%;
+  // width: 75%;
   .write-data--subscriptions-title {
     padding: $cf-space-2xs;
     display: flex;
